minimal and buggy 3d map implementation
