+++ inc/cub3D.h
@@ -6,7 +6,7 @@
 /*   By: ftroiter <ftroiter@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/24 22:54:31 by amak              #+#    #+#             */
-/*   Updated: 2024/04/08 15:24:01 by ftroiter         ###   ########.fr       */
+/*   Updated: 2024/04/08 17:01:57 by ftroiter         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,8 +46,8 @@
 # define FOV			1.047198
 # define HALF_FOV		0.523599
 # define PLANE_DIST		588.007
-# define SCREEN_WIDTH	680	
-# define SCREEN_HEIGHT	480
+# define SCREEN_WIDTH	1024
+# define SCREEN_HEIGHT  768
 
 # define MAX_TEXTURES 4
 
@@ -105,12 +105,12 @@ typedef struct s_texture
 	int			height;
 }				t_texture;
 
-typedef struct s_windows
+typedef struct s_window
 {
 	void		*mlx;
 	void		*win;
 	t_image		image;
-}				t_windows;
+}				t_window;
 
 typedef struct s_file
 {
@@ -124,8 +124,9 @@ typedef struct s_file
 	int			rows;
 	int			columns;
 	t_player	player;
-	t_windows	graphic;
+	t_window	graphic;
 	t_texture	textures[4];
+	t_ray		rays[SCREEN_WIDTH];
 }				t_file;
 
 /* EXIT FUNCTIONS */
@@ -160,7 +161,7 @@ void	init_mlx(t_file *file);
 int		key_press(int keycode, t_file *file);
 
 /* BUILD MAP FUNCTIONS*/
-void	draw_map(t_file *file, t_windows *graphic);
+void	draw_map(t_file *file, t_window *graphic);
 void	put_square(t_image *image, int x, int y, int color, int out_color);
 
 
@@ -178,7 +179,11 @@ void	increment_ray_length(t_ray *ray, float step_y, float step_x);
 /* RAYCASTING */
 void	draw_ray(t_image *image, t_player *player, t_file *file, float angle);
 void	castray(t_ray *ray, t_player *player, t_file *file, float angle);
-void	raycasting(t_file *file);
+void	raycasting(t_file *file, t_ray *rays);
+void	render_cicle(t_file *file);
+void	draw_walls(t_file *file, t_ray *rays);
+
+
