+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: ftroiter <ftroiter@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/24 22:30:11 by amak              #+#    #+#             */
-/*   Updated: 2024/04/08 01:03:12 by ftroiter         ###   ########.fr       */
+/*   Updated: 2024/04/08 17:00:27 by ftroiter         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
